Adding Homepage to get username (add other features such as get alarm times) and navigation button to Alarm page.
