+++ CPSC411_Project/Timer/TimerFunctionality.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-extension ContentView{
+extension AlarmSettingsView{
     final class ViewModel: ObservableObject{
         @Published var isRunning = false
         @Published var activeAlert = false
